reestructure main, create utils file
