+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod vm;
 pub mod condition_flag;
 pub mod opcode;
 pub mod instructions;
+pub mod utils;
