+++ src/main.rs
@@ -1,27 +1,15 @@
-use std::env;
-
+use lc3_vm::utils::*;
+use lc3_vm::vm::VM;
 fn main() {
-    let args: Vec<String> = env::args().collect();
+    // Load Arguments
+    load_arguments();
+    
+    // Setup
+    setup();
 
-    if args.len() < 2 {
-        eprintln!("lc3 [image-file1] ...");
-        return;
-    }
-
-    // Iterate over each argument (skipping the first one which is the program name)
-    for arg in &args[1..] {
-        if !read_image(arg) {
-            // Show error message and exit with status code 1
-            eprintln!("failed to load image: {}", arg);
-            return;
-        }
-    }
+    // Run VM
+    let mut vm = VM::new();
+    vm.run();
 }
 
-fn read_image(filename: &str) -> bool {
-    // Open file, read, close file. To be Implemented.
 
-    // Dummy
-    println!("Reading image: {}", filename);
-    true
-}
